requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astunparse==1.6.3
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
Cython==3.1.2
decorator==5.2.1
demucs @ git+https://github.com/facebookresearch/demucs@e976d93ecc3865e5757426930257e200846a520a
diffq==0.2.4
dora_search==0.1.12
einops==0.8.1
exceptiongroup==1.3.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
flatbuffers==1.12
fsspec==2025.7.0
future==1.0.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.74.0
h11==0.12.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
keras==2.9.0
Keras-Preprocessing==1.1.2
lameenc==1.8.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.43.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
namex==0.1.0
networkx==3.2.1
norbert==0.2.1
numba==0.60.0
numpy==1.24.4
oauthlib==3.3.1
omegaconf==2.3.0
openai-whisper==20250625
openunmix==1.3.0
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==1.5.3
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
retrying==1.4.1
rfc3986==1.5.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spleeter==2.4.0
submitit==1.5.3
sympy==1.14.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tqdm==4.67.1
treetable==0.2.5
typer==0.16.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
